Separate into components
The original TSDemuxer was a single monolithic file with several classes
defined in one place.  We want to break those classes into their own
components.

This will eventually make it easier to test and generally understand the
architecture of the demuxer.
