--- packages/hello/src/__test__/index.test.ts
@@ -1,5 +1,0 @@
-describe('Hello', () => {
-	it('World', () => {
-		expect(1).toBe(1)
-	})
-})
