--- packages/hello/src/index.ts
@@ -1,3 +1,0 @@
-export function hello(): boolean {
-	return true
-}
