+++ packages/mpegts-demuxer/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@tvkitchen/mpegts-demuxer",
+  "description": "Extracts packets from an MPEG transport stream.",
+  "version": "0.0.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/tvkitchen/utilities.git",
+    "directory": "packages/mpegts-demuxer"
+  },
+  "license": "LGPL-3.0",
+  "main": "lib/index.js",
+  "publishConfig": {
+    "access": "public"
+  }
+}
