+++ packages/mpegts-demuxer/src/__test__/MpegTsDemuxer.test.ts
@@ -1,0 +1,7 @@
+import { MpegTsDemuxer } from '../'
+
+describe('MpegTsDemuxer', () => {
+	it('should be a class', () => {
+		expect(MpegTsDemuxer).toBe()
+	})
+})
