+++ packages/mpegts-demuxer/src/classes/Payload.ts
@@ -1,0 +1,7 @@
+export type Payload = {
+	buffer: Uint8Array[];
+	buflen: number;
+	pts: number;
+	dts: number;
+	frame_ticks: number;
+};
