+++ packages/mpegts-demuxer/src/classes/index.ts
@@ -1,0 +1,4 @@
+export { Pmt } from './Pmt'
+export { Stream } from './Stream'
+export { Packet } from './Packet'
+export { Payload } from './Payload'
