+++ packages/mpegts-demuxer/src/index.ts
@@ -1,0 +1,1 @@
+export { MpegTsDemuxer } from './MpegTsDemuxer'
